packages/backend/requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
sqlalchemy==2.0.23
alembic==1.18.4
asyncpg==0.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.12.5
bcrypt==4.3.0
PyJWT==2.10.1
pgvector==0.4.2
httpx==0.28.1
pytest==9.0.2
pytest-asyncio==1.3.0
